requirements.txt
playwright
pillow
img2pdf
pikepdf
tqdm

# Testing dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

